+++ R/ggcoxdiagnostics.R
@@ -15,8 +15,8 @@
 #'@param hline.col,hline.size,hline.lty,hline.alpha,hline.yintercept color, size, linetype, visibility and Y-axis coordinate to be used for \link{geom_hline}.
 #'Used only when \code{hline = TRUE}.
 #'@param hline a logical - should the horizontal line be added to highlight the \code{Y=0} level.
-#'@param ggtheme function, ggplot2 theme name. Default value is \link{theme_classic2}.
-#'  Allowed values include ggplot2 official themes: see \link{ggtheme}.
+#'@param ggtheme function, ggplot2 theme name. Default value is ggplot2::theme_bw().
+#'  Allowed values include ggplot2 official themes: see \code{\link[ggplot2]{ggtheme}}.
 #'@param font.main,font.x,font.y,font.tickslab a vector of length 3
 #'  indicating respectively the size (e.g.: 14), the style (e.g.: "plain",
 #'  "bold", "italic", "bold.italic") and the color (e.g.: "red") of main title,
@@ -32,10 +32,11 @@
 #'@importFrom stats residuals
 #'@examples
 #'
+#'if(require(RTCGA.clinical)){
 #' # TCGA data exmaple (http://cancergenome.nih.gov/)
 #' # source("https://bioconductor.org/biocLite.R")
 #' # biocLite("RTCGA.clinical") # data for examples
-#' library(RTCGA.clinical) # also loads 'RTCGA' package
+#' # library(RTCGA.clinical) # also loads 'RTCGA' package
 #' survivalTCGA(BRCA.clinical, OV.clinical,
 #'             extract.cols = c("admin.disease_code", "patient.days_to_birth")) -> BRCAOV.survInfo
 #' BRCAOV.survInfo$age  <- round((-as.numeric(BRCAOV.survInfo$patient.days_to_birth))/365,2)
@@ -48,11 +49,11 @@
 #' ggcoxdiagnostics(coxph.fit, ggtheme = theme_RTCGA(), # library(RTCGA.clinical)
 #'  type = "deviance") + ylab('Deviance Residuals')
 #' ggcoxdiagnostics(coxph.fit, ggtheme = theme_light(), linear.predictions = FALSE)
-#' ggcoxdiagnostics(coxph.fit, ggtheme = theme_void(), type = "deviance", linear.predictions = FALSE)
-#'
-#'
+#' # ggcoxdiagnostics(coxph.fit, ggtheme = theme_void(), type = "deviance", linear.predictions = FALSE)
+#' }
 #'
 #' # traditional example
+#' library(survival)
 #' coxph.fit2 <- coxph(Surv(futime, fustat) ~ age + ecog.ps, data=ovarian)
 #' ggcoxdiagnostics(coxph.fit2, type = "deviance")
 #'
@@ -64,40 +65,51 @@ ggcoxdiagnostics <- function (fit,
                       ...,
                       linear.predictions = TRUE,
                       hline = TRUE,
-                      hline.col = "black", hline.size = 3, hline.alpha = 1, hline.yintercept = 0, hline.lty = 'solid',
+                      hline.col = "red", hline.size = 1, hline.alpha = 1, hline.yintercept = 0, hline.lty = 'dashed',
                       point.col = "black", point.size = 1, point.shape = 19, point.alpha = 1,
                       font.main = c(16, "plain", "black"),
                       font.x = c(14, "plain", "black"), font.y = c(14, "plain", "black"),
                       font.tickslab = c(12, "plain", "black"),
-                      ggtheme = theme_classic2()){
+                      ggtheme = ggplot2::theme_bw()){
 
   model <- fit
   if(!methods::is(model, "coxph"))
     stop("Can't handle an object of class ", class(fit))
+  type <- match.arg(type)
 
-  if (linear.predictions) {
-    data2plot <- data.frame(lp = predict(model, type="lp"),
-                            res = residuals(model, type = type))
-    gplot <- ggplot(aes(lp, res), data = data2plot) +
-      xlab("Linear Predictions") + ggtheme
-  } else {
-    data2plot <- data.frame(index = 1:model$n,
-                            res = residuals(model, type = type))
-    gplot <- ggplot(aes(index, res), data = data2plot) +
-      xlab("The index number of observations") + ggtheme
-  }
+  res <- as.data.frame(resid(fit, type = type))
+  .facet <- FALSE
 
-  if (hline)
-    gplot <- gplot + geom_hline(yintercept=hline.yintercept, col = hline.col,
-                                size = hline.size, lty = hline.lty,
-                                alpha = hline.alpha)
+  xlabel <- "The index number of observations"
+  ylabel <- paste0("Residuals (type = ", type, ")" )
 
+  if(linear.predictions){
+    xval <- predict(fit, type="lp")
+    xlabel <- "Linear Predictions"
+  }else xval <- 1:fit$n
+
+  # Case of multivariate Cox model
+  if(type %in% c("martingale", "deviance")) col_names <- "residuals"
+  else col_names <- names(stats::coef(fit))
+  colnames(res) <- col_names
+  res$xval <- xval
+  data2plot <- tidyr::gather_(res,
+                              key_col = "covariate", value_col = "res",
+                              gather_col = col_names)
+
+  gplot <- ggplot(aes(xval, res), data = data2plot) +
+           geom_point(col = point.col, shape = point.shape,
+                       size = point.size, alpha = point.alpha)
+
+  if (hline) gplot <- gplot + geom_hline(yintercept=hline.yintercept, col = hline.col,
+                                size = hline.size, lty = hline.lty, alpha = hline.alpha)
+
+  gplot <- gplot + labs(x = xlabel, y = ylabel) + ggtheme
   # customization
-  gplot <- gplot + geom_point(col = point.col, shape = point.shape,
-                              size = point.size, alpha = point.alpha)
   gplot <-.labs(p = gplot, font.main = font.main, font.x = font.x, font.y = font.y)
   gplot <- .set_ticks(gplot, font.tickslab = font.tickslab)
 
+  gplot <- gplot + facet_wrap(~covariate, scales = "free", ncol = 1)
   gplot
 }